Record optout stats for users with 1+ edits and 100+ edits
These have been referred to at various places so let's track them.
Don't put them in the tsv file for now, the dashboard will be
crowded already and this can be changed easily later if needed,
but there is no way to get the data later if we don't do it now.

Change-Id: Ic80e0c2bef97b28d51f69a74a809ef6c82dd0424