Fix SQL queries
Update EventLogging IDs (didn't try to be intelligent about
it and look at the dates - probably a lot of unions could be
dropped now, but I had neither time nor care), find alternative
for ImageMetrics data (which was undeployed).

Not fixed: optouts (UserDailyContribs was undeployed, don't see
any alternative source for user activity).

perf is still empty for most wikis (but not all). Not sure what
to make of that.

Bug: T98449
Change-Id: I7e10ec6ae1b94158b0d1fcb804757e99aee08756